requirements.txt
scapy
pywifi
rich
streamlit
plotly
fastapi
pandas
pyyaml
pytest
requests
